+++ HISTORY.md
@@ -1,4 +1,7 @@
-# 0.1
+# 0.2
+- support journey diagram
+- methods for using by one line
 
-1. Show gitlab pipeline as gantt diagram
-2. Show gitlab pipeline as graph LR
+# 0.1
+- Show gitlab pipeline as gantt diagram
+- Show gitlab pipeline as graph LR